added modular javascript to awesome books
