+++ index.html
@@ -14,6 +14,7 @@
       <a href="#new-book">Add a New Book</a>
       <a href="#contact">Contact Info</a>
     </nav>
+    <div class="date" id="date"></div>
     <section id="books">
     <div id="book-list">
       <ul id="books"></ul>
@@ -44,6 +45,6 @@
     <footer>
       <h2>Copyright ...</h2>
     </footer>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
   </body>
 </html>
