+++ modules/time.js
@@ -1,0 +1,7 @@
+const time = () => {
+  const timeDate = document.getElementById('date');
+  const date = new Date();
+  timeDate.innerHTML = date;
+};
+
+export default time;
